replace ChatInputBar, move logic to InputBarAccessoryView
